Big update. Now a ROS2 node.
Due to issues with getting the command line interface to operate at the
latency required for a good experience, rostui is dropping ros1 support.

This change allows me to set up the object_controller as a ros2 node,
which has access to things like 'this->get_*' which is orders of
magnitude faster. As a side effect of this, most of the monitor
functionality is deprecated, at least for now. I liked the layout of
having a dedicated class asynchronously collecting the information for
the nodes/services/topics. I can replicate this by having the monitors
spun as nodes, but I don't like having such a large footprint. Might
just need to compromise somehow.